requirements.txt
google-generativeai==0.3.1
streamlit
joblib==1.3.2
dotenv
